requirements.txt
datasets>=2.19.0
numpy>=1.26.0
pandas>=2.2.0
scikit-learn>=1.5.0
scikit-image>=0.24.0
matplotlib>=3.9.0
seaborn>=0.13.2
pillow>=10.4.0
opencv-python-headless>=4.10.0.84
torch>=2.3.0
torchvision>=0.18.0
tqdm>=4.66.0
